whew I have loading level now
but there is still problem.

you still load in your thread. and it hangs for a while in your thread.

update loading bar did not work smoothly because that loading bar also lives in this thread that are hanging loading levels.

we to convert this to  multithread version. but nobody taught it so yet. the document there are untested.

we neeed help. pls help

now we have heurestic for gameplayment loading screen. if still loading, we show next menu, otherwise show gameplay UI.

so our idea is you can pause game while loading level.

we want to build pause menu. use same main menu. that is why there is is playing game to check. pause main menu is different parameter variable but same node