Implemented the Brackeys Loading Bar...
but it's Joel version that is inspired from it.
here is demo script of it: https://github.com/Perkedel/Where_Is_LoadingBar_Functionality/blob/master/GenLoad.gd
just go to Repository main page, clone to your disk, and have fun with it. use Godot since 3.1!

But, the system still hangs a little bit when it is going to be instanced and spawned.
WHY?!?!?!?!