Aquire new file
The Particle material is limited on EEVEE. full working only on Cycles. pls fix, Blender

also, when will we subscribe to cloud.blender.org to grab all of these files??

well don't let me fall to this hype of trading! I am not having proof that it will swiftly grants money that easy. it's risky.