metaballed
we have followed YouTube tutorial how to sculpt in Blender.

one of the way is to use Metaball base building.

basically build the sculpt clay rough out of metaballs.

also there is another way by using edge lines and then Skin modifier it to build sculpt base.

I think Metaball is best for both process either body and limb parts for now.

the edge line method is cumbersome to handle. sometimes glitch can happens and may ruins the build, as well as steal alot of attention time. This is very dangerous for especially working for client a.k.a. commission. But not limited to for regular pace job that made for self and hobby.